Changed over a bunch of FBX files. see description for something about canvas
Inside of the prefab for stage 3, it wasn't allowing me to save as there was a canvas object, with a bunch of inventory details inside of it which had no references to anything. Like every single piece of that inventory was broken, so I just deleted it. But I believe the actual inventory canvas is still working.